+++ packages/util/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @workglow/util
 
+## 0.3.44
+
+### Features
+
+#### util
+
+- declare x-ui-placeholder and x-ui-disabled
+
+## 0.3.43
+
+## 0.3.42
+
+## 0.3.41
+
+## 0.3.40
+
 ## 0.3.39
 
 ### Features